+++ src/app/components/forms/terra-forms.example.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { TerraSelectBoxValueInterface } from './select-box/data/terra-select-box.interface';
+import { TerraMultiCheckBoxValueInterface } from './multi-check-box/data/terra-multi-check-box-value.interface';
 
 @Component({
     selector: 'terra-forms-example',
@@ -23,6 +24,35 @@ export class TerraFormsExampleComponent
             value:   2
         }
     ];
+
+    protected readonly multiCheckboxValues:Array<TerraMultiCheckBoxValueInterface> =
+        [
+            {
+                value:    '1',
+                caption:  'Value 1',
+                selected: false
+            },
+            {
+                value:    '2',
+                caption:  'Value 2',
+                selected: true
+            },
+            {
+                value:    '3',
+                caption:  'Value 3',
+                selected: true
+            },
+            {
+                value:    '4',
+                caption:  'Value 4',
+                selected: false
+            },
+            {
+                value:    '5',
+                caption:  'Value 5',
+                selected: false
+            }
+        ];
 
     protected submit():void
     {