+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 * **terra-slider** implement `ControlValueAccessor` interface to be able to use ngModel instead of `inputValue`.
 * **terra-select-box** is now focusable while navigating with tab key
+* **terra-check-box** is now focusable while navigating with tab key
 * **terra-radio-input** new component that wraps html's native input of type radio including it's label. It must be used within a `<terra-radio-group>` element.
 * **terra-radio-group** new component to contain and manage `<terra-radio-input>` elements. It implements the `ControlValueAccessor` interface (`ngModel`) to be able to use it in a form.
 * **terra-radio-button** is now deprecated. Use `<terra-radio-input>` and `<terra-radio-group>` instead.